requirements.txt
Flask==0.12
Jinja2==2.9.6
MarkupSafe==1.0
Werkzeug==0.12.2
argparse==1.2.1
click==6.7
dnspython==1.15.0
itsdangerous==0.24
wsgiref==0.1.2
